byodforecast-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.0
boto3==1.29.3
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
plotly==5.18.0
optuna==3.5.0
lightgbm==4.1.0 